Removed the picklefield hack.
Was previously storing next dates essentially as a comma separated list
and converting it back and forth to a array when I needed it. Yeah.
So I put on my schema-rewriting pants and made a new model with a
many-to-many connection to foods for upcoming dates and a one-to-many
field for the last date. Updated tests, tasks & admin accordingly and used
South to migrate. #killingit